Allow alarms on only specific days - fix some minor bugs
This  set of changes allows the program to set alarms which only work
on specific days of the week.  By default, they all happen on all days,
until you turn some days off.  This also adds the menu item to allow
the user to re-do the last alarm.  You can then snooze it.  (Useful if
you accidentally ack'ed an alarm which you didn't mean to.)  It also
changes the default snooze time from 10 to 20 minutes.